Optimized by OpenMP
Use OpenMP to create multiple threads to execute loops in raytracing function.
The execution time of the OpenMP version is about half of the original
if number of threads is two or more.
Number of threads can be adjusted by assigning Makefile target
when using make to construct a program.